ai-learning-phase2/pyproject.toml
[project]
name = "ai-learning-phase2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.14",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "tiktoken>=0.12.0",
    "torch>=2.11.0",
]
